[3D] Slightly reduced memory footprint: Use uint16 instead of uint32 for indices in meshes

+++ Rendering/TheEngine/Interfaces/IMesh.cs
@@ -7,7 +7,7 @@ namespace TheEngine.Interfaces
 {
     public interface IMesh
     {
-        void SetIndices(ReadOnlySpan<uint> indices, int submesh);
+        void SetIndices(ReadOnlySpan<ushort> indices, int submesh);
         void SetSubmeshIndicesRange(int submesh, int start, int length);
         void Rebuild();
         void Activate();
