+++ Rendering/TheEngine/Primitives/ScreenPlane.cs
@@ -28,7 +28,7 @@ namespace TheEngine.Primitives
         };
 
 
-        private static uint[] indices =
+        private static ushort[] indices =
         {
             0, 1, 2,
             0, 2, 3
